Update command.py
It is possible for _poll_check_status to be calling an instance with only a partial allocation of slots. Redis 3.0.5 also allows slaves of one master to be re-assigned to a new master via CLUSTER FAILOVER, so there is no need to ensure the slave is not connected to an existing cluster.